Changes to get the Pojo acceptance test running.  Works in IntelliJ but not the command line.

+++ driver/src/test/functional/org/mongodb/DatabaseTestCase.java
@@ -47,7 +47,9 @@ public class DatabaseTestCase {
 
     @After
     public void tearDown() {
-        collection.tools().drop();
+        if (collection != null) {
+            collection.tools().drop();
+        }
     }
 
     protected String getDatabaseName() {
